refactor(clients): Refactor Clients to have separate clients for Movements and Transfers
