+++ spec/vcr/Fintoc_Clients_MovementsClient/_get_link/get_the_link_from_a_given_link_token.yml
