+++ spec/vcr/Fintoc_Clients_MovementsClient/_link/get_the_link_from_a_given_link_token.yml
